Move SmartContractSaveCoinviewRule to own file
