+++ src/Stratis.Bitcoin.Features.SmartContracts/Consensus/Rules/SmartContractLoadCoinviewRule.cs
@@ -3,7 +3,6 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using NBitcoin;
-using Stratis.Bitcoin.Base.Deployments;
 using Stratis.Bitcoin.Consensus.Rules;
 using Stratis.Bitcoin.Features.Consensus;
 using Stratis.Bitcoin.Features.Consensus.CoinViews;
@@ -12,23 +11,6 @@ using Stratis.Bitcoin.Utilities;
 
 namespace Stratis.Bitcoin.Features.SmartContracts.Consensus.Rules
 {
-    [FullValidationRule]
-    public class SmartContractSaveCoinviewRule : UtxoStoreConsensusRule
-    {
-        /// <inheritdoc />
-        public override async Task RunAsync(RuleContext context)
-        {
-            uint256 oldBlockHash = context.ValidationContext.ChainTipToExtand.Previous.HashBlock;
-            uint256 nextBlockHash = context.ValidationContext.ChainTipToExtand.HashBlock;
-
-            // Persist the changes to the coinview. This will likely only be stored in memory,
-            // unless the coinview treashold is reached.
-            this.Logger.LogTrace("Saving coinview changes.");
-            var utxoRuleContext = context as UtxoRuleContext;
-            await this.PowParent.UtxoSet.SaveChangesAsync(utxoRuleContext.UnspentOutputSet.GetCoins(this.PowParent.UtxoSet), null, oldBlockHash, nextBlockHash).ConfigureAwait(false);
-        }
-    }
-
     [PartialValidationRule]
     public sealed class SmartContractLoadCoinviewRule : UtxoStoreConsensusRule
     {
